Fix Workflow loading issue
There was an issue when loading a Workflow containing a KPI which uses an available variable as input (e.g. with a Workflow where there is an MCO parameter "V1" and a KPI calculator which uses "V1" as input). The issue was that when creating the KPI calculator model view, the input variable in the KPI calculator model wasn't in the available variables yet, resulting in a trait error because the input variable was not available.

I could fix this issue by using the post_init option of on_trait_change, which ensure that the WorkflowModelView and the available variables are correctly updated.