Fix issue when changing the size of the input/output slots
When emptying the input_slot_representation list by replacing it with [], a reference to the elements contained in the list was still there. Resulting in an index error when changing the input_slot_maps of the model. I could fix it by emptying the content of the list: input_slot_representation[:] = [] instead.